# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/messages": {
            "get": {
                "description": "Get messages by sent status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Get messages by sent status",
                "parameters": [
                    {
                        "enum": [
                            "pending",
                            "sent",
                            "failed"
                        ],
                        "type": "string",
                        "description": "Sent status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/message.GetMessagesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/message.GetMessagesResponse"
                        }
                    }
                }
            }
        },
        "/messages/job-state": {
            "post": {
                "description": "Start or stop the message sender job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job"
                ],
                "summary": "Change the state of the message sender job",
                "parameters": [
                    {
                        "enum": [
                            "start",
                            "stop"
                        ],
                        "type": "string",
                        "description": "State to change to",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/messages/queue": {
            "post": {
                "description": "Queue a message for sending",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Queue a message for sending",
                "parameters": [
                    {
                        "description": "Create form to be queued",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/message.CreateMessageForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "message.CreateMessageForm": {
            "type": "object",
            "required": [
                "content",
                "recipient_phone"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "recipient_phone": {
                    "type": "string"
                }
            }
        },
        "message.GetMessagesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Message"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Message": {
            "type": "object",
            "required": [
                "content",
                "recipient_phone",
                "sent_status"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "recipient_phone": {
                    "type": "string"
                },
                "remote_id": {
                    "type": "string"
                },
                "sent_at": {
                    "type": "string"
                },
                "sent_status": {
                    "$ref": "#/definitions/models.SentStatus"
                }
            }
        },
        "models.SentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "sent",
                "failed"
            ],
            "x-enum-varnames": [
                "SentStatusPending",
                "SentStatusSent",
                "SentStatusFailed"
            ]
        },
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}